move all functions to their own files
