+++ src/index.ts
@@ -1,4 +1,7 @@
-export {TemplateResult, html, render} from './html.js'
+export {TemplateResult} from './template-result.js'
+export {render} from './render.js'
+export {processor} from './processor.js'
+export {html} from './html.js'
 export {isDirective, directive} from './directive.js'
 export {until} from './until.js'
 export {unsafeHTML} from './unsafe-html.js'
